+++ CHANGELOG.md
@@ -8,6 +8,10 @@ for the globale releases (`YYYY.MINOR`), and to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html) for `libnopegl`.
 
 ## [Unreleased]
+### Added
+- `ngl_scene_get_filepaths()` and `ngl_scene_update_filepath()` to introspect
+  and manipulate the filepath based parameters in the scene graph
+
 ### Changed
 - `Text.font_files` text-based parameter is replaced with `Text.font_faces` node
   list which accepts `FontFace` nodes instead